Refactor bot runtime into modular services
