+++ src/index.ts
@@ -3,8 +3,27 @@ export { BuilderService } from './builder/builder.service';
 export {
     BotRuntime,
     IBotRuntimeOptions,
+    BotRuntimeDependencies,
     normalizeBotOptions,
 } from './builder/bot-runtime';
+export {
+    PageNavigator,
+    PageNavigatorFactoryOptions,
+    IValidationResult,
+    createPageNavigator,
+} from './builder/runtime/page-navigator';
+export {
+    SessionManager,
+    SessionManagerFactoryOptions,
+    IChatSessionState,
+    createSessionManager,
+} from './builder/runtime/session-manager';
+export {
+    PersistenceGateway,
+    PersistenceGatewayFactoryOptions,
+    IContextDatabaseState,
+    createPersistenceGateway,
+} from './builder/runtime/persistence-gateway';
 export {
     BOT_BUILDER_MODULE_OPTIONS,
     BOT_BUILDER_PRISMA,
